Changes to make system more stablewq

+++ Architecture/Controllers/MapBuilder.c
@@ -58,21 +58,15 @@ static XYCoordinate_t PopulateGridsAndUpdatePosition(
                 i,
                 ypos);
 
-//            if(DistanceSensor_GetDistanceInCm(instance->leftSensor) < DistanceToCountAWallInThatDirectionCm)
-//            {
-                GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
-                    blockedAreasGrid,
-                    i,
-                    ypos + 1);
-//            }
+            GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
+                blockedAreasGrid,
+                i,
+                ypos + 1);
 
-//            if(DistanceSensor_GetDistanceInCm(instance->rightSensor) < DistanceToCountAWallInThatDirectionCm)
-//           {
-                GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
-                    blockedAreasGrid,
-                    i,
-                    ypos - 1);
-//           }
+            GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
+                blockedAreasGrid,
+                i,
+                (int8_t)(ypos - 1));
         }
 
         xpos += numberOfGridCellsToOccupy;
@@ -88,21 +82,15 @@ static XYCoordinate_t PopulateGridsAndUpdatePosition(
                 i,
                 ypos);
 
-//            if(DistanceSensor_GetDistanceInCm(instance->leftSensor) < DistanceToCountAWallInThatDirectionCm)
-//            {
-                GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
-                    blockedAreasGrid,
-                    i,
-                    ypos - 1);
-//            }
+            GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
+                blockedAreasGrid,
+                i,
+                (int8_t)(ypos - 1));
 
-//            if(DistanceSensor_GetDistanceInCm(instance->rightSensor) < DistanceToCountAWallInThatDirectionCm)
-//           {
-                GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
-                    blockedAreasGrid,
-                    i,
-                    ypos + 1);
-//           }
+            GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
+                blockedAreasGrid,
+                i,
+                ypos + 1);
         }
 
         xpos -= numberOfGridCellsToOccupy;
@@ -118,21 +106,15 @@ static XYCoordinate_t PopulateGridsAndUpdatePosition(
                 xpos,
                 i);
 
-//            if(DistanceSensor_GetDistanceInCm(instance->leftSensor) < DistanceToCountAWallInThatDirectionCm)
-//            {
-                GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
-                    blockedAreasGrid,
-                    xpos - 1,
-                    i);
-//            }
+            GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
+                blockedAreasGrid,
+                (int8_t)(xpos - 1),
+                i);
 
-//            if(DistanceSensor_GetDistanceInCm(instance->rightSensor) < DistanceToCountAWallInThatDirectionCm)
-//           {
-               GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
-                   blockedAreasGrid,
-                   xpos + 1,
-                   i);
-//           }
+           GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
+               blockedAreasGrid,
+               xpos + 1,
+               i);
         }
 
         ypos += numberOfGridCellsToOccupy;
@@ -148,21 +130,15 @@ static XYCoordinate_t PopulateGridsAndUpdatePosition(
                 xpos,
                 i);
 
-//            if(DistanceSensor_GetDistanceInCm(instance->leftSensor) < DistanceToCountAWallInThatDirectionCm)
-//            {
-                GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
-                    blockedAreasGrid,
-                    xpos + 1,
-                    i);
-//            }
+            GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
+                blockedAreasGrid,
+                xpos + 1,
+                i);
 
-//            if(DistanceSensor_GetDistanceInCm(instance->rightSensor) < DistanceToCountAWallInThatDirectionCm)
-//           {
-               GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
-                   blockedAreasGrid,
-                   xpos - 1,
-                   i);
-//           }
+           GridMap_FirstQuadrant5cmCell3m2x3m2_SetCellValueAtIndex(
+               blockedAreasGrid,
+               (int8_t)(xpos - 1),
+               i);
         }
 
         ypos -= numberOfGridCellsToOccupy;
